new model component has been created for data interfaces.
